Refactor HistogramPanelFX into HistogramChart
Aim to make histograms easier to create and display.
Current implementation breaks a bit with the 'series' concept; data should be added through the list returned by `getHistogramData()`. This simplifies the implementation (by simply subclassing AreaChart).
